+++ packages/acceptance-tests/pkg-tests-specs/sources/pnp-esm.test.ts
@@ -499,4 +499,64 @@ describe(`Plug'n'Play - ESM`, () => {
       },
     ),
   );
+
+  test(
+    `it should throw ERR_REQUIRE_ESM when requiring a file with type=module`,
+    makeTemporaryEnv(
+      {
+        dependencies: {
+          'no-deps-esm': `1.0.0`,
+        },
+      },
+      {
+        pnpEnableEsmLoader: true,
+      },
+      async ({path, run, source}) => {
+        await expect(run(`install`)).resolves.toMatchObject({code: 0});
+
+        await xfs.writeFilePromise(ppath.join(path, `index.js` as Filename), `
+          try {
+            require('no-deps-esm')
+          } catch (err) {
+            console.log(err.code)
+          }
+        `);
+
+        await expect(run(`node`, `index.js`)).resolves.toMatchObject({
+          code: 0,
+          stdout: `ERR_REQUIRE_ESM\n`,
+        });
+      },
+    ),
+  );
+
+  test(
+    `it should throw ERR_REQUIRE_ESM when requiring a .mjs file`,
+    makeTemporaryEnv(
+      {
+        dependencies: {
+          'no-deps-mjs': `1.0.0`,
+        },
+      },
+      {
+        pnpEnableEsmLoader: true,
+      },
+      async ({path, run, source}) => {
+        await expect(run(`install`)).resolves.toMatchObject({code: 0});
+
+        await xfs.writeFilePromise(ppath.join(path, `index.js` as Filename), `
+          try {
+            require('no-deps-mjs/index.mjs')
+          } catch (err) {
+            console.log(err.code)
+          }
+        `);
+
+        await expect(run(`node`, `index.js`)).resolves.toMatchObject({
+          code: 0,
+          stdout: `ERR_REQUIRE_ESM\n`,
+        });
+      },
+    ),
+  );
 });